+++ neovim_lua/plugins/neo-tree.lua
@@ -16,6 +16,6 @@ return {
                 },
             },
         })
-        vim.keymap.set("n", "<C-n>", ":Neotree filesystem reveal left<CR>")
+        vim.keymap.set("n", "<C-n>", ":Neotree filesystem reveal left<CR>", { desc = "Neo-tree: reveal file" })
     end,
 }