+++ neovim_lua/plugins/telescope.lua
@@ -1,14 +1,14 @@
 return {
     {"nvim-telescope/telescope.nvim",
-        tag = "0.1.5",
+        version = "^0.2.0",
         dependencies = {
             'nvim-lua/plenary.nvim'
         },
         lazy = false,
         config = function()
             local builtin = require("telescope.builtin")
-            vim.keymap.set('n', '<C-p>', builtin.find_files, {})
-            vim.keymap.set('n', '<leader>fg', builtin.live_grep, {})
+            vim.keymap.set('n', '<C-p>', builtin.find_files, { desc = "Telescope: find files" })
+            vim.keymap.set('n', '<leader>fg', builtin.live_grep, { desc = "Telescope: live grep" })
         end
     },
     {"nvim-telescope/telescope-ui-select.nvim",